+++ R/iFitbit.R
@@ -37,7 +37,6 @@
 #' @importFrom utils modifyList tail menu
 #' @importFrom rmarkdown render
 #' @importFrom dplyr distinct case_when summarise_at select desc mutate_if top_n mutate arrange rename bind_cols if_else summarise_all
-#' @importFrom magick image_read
 #' @importFrom reactable colDef colGroup reactable reactableTheme
 #' @importFrom purrr as_vector
 #' @importFrom tidyr pivot_longer complete
@@ -47,4 +46,5 @@
 #' @importFrom graphics layout
 #' @importFrom glue glue glue_collapse
 #' @importFrom rlang inform
+#' @importFrom lubridate force_tz
 NULL